+++ src/types/mod.rs
@@ -32,29 +32,29 @@ pub const MAX_TITLE_SIZE: EN_SizeLimits = 79;
 #[derive(Debug, Copy, Clone, PartialEq, FromPrimitive)]
 #[repr(i32)]
 pub enum ObjectType {
-    Node = EN_ObjectType_EN_NODE, // Nodes
-    Link = EN_ObjectType_EN_LINK, // Links
-    TimePattern = EN_ObjectType_EN_TIMEPAT, // Time patterns
-    Curve = EN_ObjectType_EN_CURVE, // Data curves
-    Control = EN_ObjectType_EN_CONTROL, // Simple controls
-    Rule = EN_ObjectType_EN_RULE, // Control rules
+    Node = EN_ObjectType_EN_NODE as i32, // Nodes
+    Link = EN_ObjectType_EN_LINK as i32, // Links
+    TimePattern = EN_ObjectType_EN_TIMEPAT as i32, // Time patterns
+    Curve = EN_ObjectType_EN_CURVE as i32, // Data curves
+    Control = EN_ObjectType_EN_CONTROL as i32, // Simple controls
+    Rule = EN_ObjectType_EN_RULE as i32, // Control rules
 }
 
 #[derive(Debug, Copy, Clone, PartialEq, FromPrimitive)]
 #[repr(i32)]
 pub enum CountType {
-    NodeCount = EN_CountType_EN_NODECOUNT, // Number of nodes (junctions + tanks + reservoirs)
-    TankCount = EN_CountType_EN_TANKCOUNT, // Number of tanks and reservoirs
-    LinkCount = EN_CountType_EN_LINKCOUNT, // Number of links (pipes + pumps + valves)
-    PatternCount = EN_CountType_EN_PATCOUNT, // Number of time patterns
-    CurveCount = EN_CountType_EN_CURVECOUNT, // Number of data curves
-    ControlCount = EN_CountType_EN_CONTROLCOUNT, // Number of simple controls
-    RuleCount = EN_CountType_EN_RULECOUNT, // Number of rule-based controls
+    NodeCount = EN_CountType_EN_NODECOUNT as i32, // Number of nodes (junctions + tanks + reservoirs)
+    TankCount = EN_CountType_EN_TANKCOUNT as i32, // Number of tanks and reservoirs
+    LinkCount = EN_CountType_EN_LINKCOUNT as i32, // Number of links (pipes + pumps + valves)
+    PatternCount = EN_CountType_EN_PATCOUNT as i32, // Number of time patterns
+    CurveCount = EN_CountType_EN_CURVECOUNT as i32, // Number of data curves
+    ControlCount = EN_CountType_EN_CONTROLCOUNT as i32, // Number of simple controls
+    RuleCount = EN_CountType_EN_RULECOUNT as i32, // Number of rule-based controls
 }
 
 #[derive(Debug, Copy, Clone, PartialEq, FromPrimitive)]
 #[repr(i32)]
 pub enum ActionCodeType {
-    Unconditional = EN_ActionCodeType_EN_UNCONDITIONAL, // Delete all controls and connecting links
-    Conditional = EN_ActionCodeType_EN_CONDITIONAL, // Cancel object deletion if it appears in controls or has connecting links
+    Unconditional = EN_ActionCodeType_EN_UNCONDITIONAL as i32, // Delete all controls and connecting links
+    Conditional = EN_ActionCodeType_EN_CONDITIONAL as i32, // Cancel object deletion if it appears in controls or has connecting links
 }
